feat: improve app navbar design and layout
